Improve publish logic
Description
===========

This patch improves the `HockeyUploadTask` in two ways.

* switch to `apache:httpclient` for upload
* save upload result metadata `HockeyAppVersion`

I moved from using curl to `httpclient`. The setup is very simple if a
little bit verbose. But it supports everything we need to `POST` a
multipart payload. The main reason why I wanted to switch is to read the
response body of the request. Hockey returns a JSON body with
informations about the created app version

**example from [support.hockeyapp.net][api-versions]
```json
{
    "title": "HockeyTest",
    "appsize": 1598428,
    "timestamp": 1308930206,
    "device_family": "iPhone/iPod",
    "minimum_os_version": "4.0",
    "notes": "<p>Some new features and fixed bugs.</p>",
    "version": "8",
    "shortversion": "1.0",
    "status": 2,
    "config_url": "https://rink.hockeyapp.net/manage/apps/123/app_versions/8",
    "public_url": "https://rink.hockeyapp.net/apps/1234567890abcdef1234567890abcdef"
}
```

> The actual body contains more fields.

The `HockeyUploadTask` will save this result as an `Outputfile` and also
parse it into a simple model `HockeyAppVersion`. This allows other tasks
to access this information.

Changes
=======

![ADD] new compile dependency `org.apache.httpcomponents:httpclient`
![ADD] new compile dependency `org.apache.httpcomponents:httpmime`
![IMPROVE] switch upload logic from `curl` to `httpclient`
![IMPORVE] save upload result body in file
![IMPROVE] provide runtime model for upload result

[api-versions]:    https://support.hockeyapp.net/kb/api/api-versions#upload-version