Switch to Provider API
Description
===========

The current plugin source used the old and internal convention mapping
logic. It seems it is still supported up to gradle `6.4`. But
understanding and using this API is not straight forward. This patch
moves to the `Provider` API introduced in gradle `4.4`. We will lose
support for some earlier gradle `4.x` versions but the code becomes
easier to read and maintain. I adjusted the `gradle_check_versions.sh`
file to test all gradle releases up to `6.4`. All worked well except
some internal defaults when it comes to work with the `Provider` API and
some breaking changes with gradle `6.x`. I fixed those with a `compat`
helper class and resolve the API break at runtime. Since the goal is to
move to gradle `6.x` by the end of the year I see no major issue with
this.

Changes
=======

* ![IMPROVE] switch to `Provider` API