Add gradle extension
Description
============

This patch adds and configures an extension point to the appCenter
plugin `AppCenterPluginExtension`. This extension allows to set the main
properties like `owner` `api token` or distribution groups.

The extension is mapped to `appCenter`

```groovy

appCenter {
   apiToken = ""
   owner = ""
   applicationIdentifier = ""
   defaultDestinations = ["", ""]
}
```

This addition is usefull if a project needs to publish multiple apks or
configures multiple upload tasks for other reasons and need to share the
base configuration.

The properties are by default looked up in the gradle properties or
environment.

| property              | gradle property name              | environment variable                |
| --------------------- | --------------------------------- | ----------------------------------- |
| apiToken              | `appCenter.apiToken`              | `APP_CENTER_API_TOKEN`              |
| owner                 | `appCenter.owner`                 | `APP_CENTER_OWNER`                  |
| applicationIdentifier | `appCenter.applicationIdentifier` | `APP_CENTER_APPLICATION_IDENTIFIER` |
| defaultDestinations   | `appCenter.defaultDestinations`   | `APP_CENTER_DEFAULT_DESTINATIONS`   |

The compatibility list had to be adjusted once more since I found out
that not all API calls of the Provider API are fully supported. The
plugin itself functions correctly. I adjusted the README

Changes
=======

* ![ADD] plugin extension `AppCenterPluginExtension`