Add publish enable flag in appcenter extension
Description
===========

This patch adds a new property `publishEnabled` to the appcenter
extensions. This property controls if the `publishAppCenter` task
created by the plugin is added as a dependency of the `publish` task.

The task can still be called indipendently. Other tasks of type
`AppCenterUploadTask` are not by default added to the `publish`
lifecycle task.

Changes
=======

* ![ADD] `publishEnable` property