Fix missing multiple value setter
Description
===========

This patch brings back the setter which allowed to set distribution
groups with `String[]` values. One native method returns these if you
call `split()` on a `String`. One could easily call `toList` on the
resulting array but this plugin had the functionality before and it was
forgotten when moving to the Provider API.

Changes
=======

* ![FIX] missing multiple value setter