Add retry logic for upload call (#26)
Description
===========

It happens quite often that the binary upload call fails with an error
in the 500 range. I have no real idea what the root cause is since it's
a server error. Maybe the inital `createUploadResource` resource call is
to blame. This patch adds a simple retry logic around the upload task to
check if our errors go down when we try to redo the call a few times
with a timeout in between.

The values for retryCount and retryTimeout can be passed as usual via
properties, env or directly in script.

| property              | gradle property name              | environment variable                |
| --------------------- | --------------------------------- | ----------------------------------- |
| retryCount            | `appCenter.retryCount`            | `APP_CENTER_RETRY_COUNT`            |
| retryTimeout          | `appCenter.retryTimeout`          | `APP_CENTER_RETRY_TIMEOUT`          |

The default values are `3` for the `retryCount` and `5` seconds for the
`retryTimeout`.

To test the logic I had to split the actual call into a helper method.
It would be cleaner to also move the other methods over but I treat this
change as a temp fix than a proper implemenation. Maybe I roll this back
...

Changes
=======

* ![ADD] retry logic for upload call