Fix deprecated publish flow
Description
===========

Appcenter suddenly without warning deprecated the old publish flow which
used a hockeyapp proxy for binary uploads. We suddenly saw 500 server
errors for the initail call to create an upload resource.
This was a planned [maintenance][appcenter_maintenance] from appcenter
which I had no knowledge of. The solution from appcener is to use the
latest version of the appcenter cli or update the upload flow based on
the cli implementation. I used the [appcenter fastlane plugin] as
reference as this was easier to read.

The new flow is somewhat similar but way more complicated.

Here are the old steps:

* create upload resource
* upload binary to provided upload url from created upload resource
  (`hockey upload url`)
* commit binary upload to the upload resource (this returns a releaseId)
* distribute release to destinations

The new flow

* create a release upload
* upload the binary to files.appcenter.ms in multiple steps
  * set binary metadata (filename, filesize). this call returns a
    chunksize
  * uplaod binary in chunks (~4MB)
  * mark upload as finished
* set state of release upload to `uploadFinished`
* poll appcenter for `releaseId` for given `uploadId`
* load the release to get download and install url
* ditribute release to destinations

I implemented these steps as static methods in `AppCenterRest` this is
something we need to address. This scope of this little plugin which
started as a simple hockeyapp upload task with a single http call needs
now somewhat more structure.

The recall logic from the last patch (#26) is partially removed. I still
kept the gradle properties etc but these are unsused in this first
implementation. The [fastlane plugin][fastlane_appcenter_helper] only
does a retry for > 500 and < 200 calls.

The fastlane plugin is also doing some release metadata upload which we
don't do at all. I would check if we need this as well or not.

Changes
=======

* ![FIX] deprecated publish flow

[appcenter_maintenance]: https://status.appcenter.ms/incidents/4nbhlm0lxsvt
[appcenter fastlane plugin]: https://github.com/microsoft/fastlane-plugin-appcenter/blob/2b8aef1aa9a75faf09f14428d50d65d9d35b34e7/lib/fastlane/plugin/appcenter/actions/appcenter_upload_action.rb#L129-L275
[fastlane_appcenter_helper]: https://github.com/microsoft/fastlane-plugin-appcenter/blob/2b8aef1aa9a75faf09f14428d50d65d9d35b34e7/lib/fastlane/plugin/appcenter/helper/appcenter_helper.rb#L313-L369