Improve rate limit handling (#33)
Description
===========

This patch adds a special handler for the `429` http response code. The
[microsoft docs] describe some headers that will be available when the
rate limit nears its end. I deciced to implement a simple handler for
the `Retry-After` header. This should contain a time value in seconds
when to retry the service again. I implemented the handler with a
generic method. The caller has to provide a closure which will be called
after the timeout. I setup a fallback timeout of 2 minutes in case the
header can't be found.

Changes
=======

* ![IMPROVE] rate limit handling

[microsoft docs]: https://docs.microsoft.com/en-us/azure/devops/integrate/concepts/rate-limits?view=azure-devops#api-client-experience