Implement simple retry logic in case of app extracting error
Description
===========

Appcenter returns an error every now and then `A problem occured while extracting your app`.
This seems to be an internal issue as the support was not yet able
to help out with this error. Under normal condidtion we trigger a rebuild
and the upload works again. This little retry addition just checks for this
specific error and retries the whole upload up to 3 times.

Changes
=======

* ![IMPROVE] uploader with retry when app extraction error is thrown