Improve upload stability by retry upload for any server error
Description
===========

This is my latest attempt to make appcenter more stable. The
developers still fight with a list off issues which results in
random server errors. I will now add another exception to
the retry worthy cases. I only do this for the `releaseID poll` call
since this is the one we always see errors and I don't want to
rewrite the whole plugin. If this will also not yield better
stability I'm inclined to wrap the whole upload in a try catch
and repeat each upload multiple times no matter the error.

Changes
=======

* ![IMPROVE] upload stability by retry upload for any server error